+++ packages/nano-ui/lib/NanoUI/Monad.hs
@@ -43,7 +43,6 @@ module NanoUI.Monad
   , styled
   , themed
   , disabledWhen
-  , followSystemThemeUi
   , systemAppearance
 
     -- * Focus and clipboard