+++ packages/nano-ui/lib/NanoUI/Testing.hs
@@ -139,6 +139,8 @@ module NanoUI.Testing
   , wrapTextLinesIO
   , caretX
   , selectionSpans
+  , textNodeFontVariant
+  , textNodeFontTone
   , textNodeFontWeight
   , textNodeFontStyle
   , textNodeTextDecoration
@@ -158,6 +160,6 @@ import NanoUI.Internal.Frame.Scroll (ScrollBarLayout (..), scrollBarLayout)
 import NanoUI.Internal.Frame.TextArea (TextInputArea (..), textInputArea)
 import NanoUI.Internal.Layout.Solve (computePopupPosition)
 import NanoUI.Internal.Monad (Ui, askContext, askHost, askInput, uiIO)
-import NanoUI.Internal.WidgetText (textNodeFontStyle, textNodeFontWeight, textNodeTextDecoration)
+import NanoUI.Internal.WidgetText (textNodeFontStyle, textNodeFontTone, textNodeFontVariant, textNodeFontWeight, textNodeTextDecoration)
 import NanoUI.Internal.Types (Damage (..), damageIsEmpty)
 import Effectful (Eff, IOE, runEff, type (:>))